requirements.txt
flask 
youtube-search-python
youtube-search-python
pytube
numpy
pandas
pymysql
cryptography
python-dotenv
flask_mysqldb